Use SVG instead of PNG for the background image
There is still a PNG fallback for browsers not supporting SVG (IE8).
This fallback is the old PNG before the increased image size, except the
file size has been trimmed down using optipng.

Note: to actually decrease the file size, compression has to be enabled
in the server for the MIME type "image/svg+xml". Otherwise the file size
only increases for browsers except IE8 (which has the fallback).